put functions and classes related to loss calculation into network class defintion
