+++ detection_vis_backend/networks/darod.py
@@ -470,14 +470,14 @@ class Decoder(nn.Module):
         return final_bboxes.detach(), final_labels.detach(), final_scores.detach()
 
 
-def roi_delta(roi_bboxes, gt_boxes, gt_labels, model_config, seed):
-    total_labels = model_config["n_class"]
-    total_pos_bboxes = int(model_config["fastrcnn"]["frcnn_boxes"] / 3)
-    total_neg_bboxes = int(model_config["fastrcnn"]["frcnn_boxes"] * 2 / 3)
+def roi_delta(roi_bboxes, gt_boxes, gt_labels, n_class, fastrcnn_cfg, seed):
+    total_labels = n_class
+    total_pos_bboxes = int(fastrcnn_cfg["frcnn_boxes"] / 3)
+    total_neg_bboxes = int(fastrcnn_cfg["frcnn_boxes"] * 2 / 3)
     device = roi_bboxes.device
-    variances = torch.tensor(model_config["fastrcnn"]["variances_boxes"]).to(device)
-    adaptive_ratio = model_config["fastrcnn"]["adaptive_ratio"]
-    positive_th = model_config["fastrcnn"]["positive_th"]
+    variances = torch.tensor(fastrcnn_cfg["variances_boxes"]).to(device)
+    adaptive_ratio = fastrcnn_cfg["adaptive_ratio"]
+    positive_th = fastrcnn_cfg["positive_th"]
 
     batch_size, total_bboxes = roi_bboxes.size(0), roi_bboxes.size(1)
     # Calculate iou values between each bbox and ground truth boxes
@@ -506,9 +506,9 @@ def roi_delta(roi_bboxes, gt_boxes, gt_labels, model_config, seed):
     roi_bbox_deltas = scatter_indices * roi_bbox_deltas.unsqueeze(-2)
     roi_bbox_deltas = roi_bbox_deltas.reshape(batch_size, total_bboxes * total_labels, 4)
 
-    if model_config["fastrcnn"]["reg_loss"] == "sl1":
+    if fastrcnn_cfg["reg_loss"] == "sl1":
         return roi_bbox_deltas.detach(), roi_bbox_labels.detach()
-    elif model_config["fastrcnn"]["reg_loss"] == "giou":
+    elif fastrcnn_cfg["reg_loss"] == "giou":
         expanded_roi_boxes = scatter_indices * roi_bboxes.unsqueeze(-2)
         expanded_roi_boxes = expanded_roi_boxes.reshape(batch_size, total_bboxes * total_labels, 4)
 
@@ -520,7 +520,7 @@ def roi_delta(roi_bboxes, gt_boxes, gt_labels, model_config, seed):
         raise ValueError("Unsupported loss type.")
 
 
-def calculate_rpn_actual_outputs(anchors, gt_boxes, gt_labels, config, seed):
+def calculate_rpn_actual_outputs(anchors, gt_boxes, gt_labels, rpn_cfg, feature_map_shape, seed):
     """
     Generating one step data for training or inference. Batch operations supported.
     :param anchors: (total_anchors, [y1, x1, y2, x2]) these values in normalized format between [0, 1]
@@ -533,13 +533,13 @@ def calculate_rpn_actual_outputs(anchors, gt_boxes, gt_labels, config, seed):
     
     batch_size = gt_boxes.size(0)
     device = gt_boxes.device
-    anchor_count = config["rpn"]["anchor_count"]
-    total_pos_bboxes = int(config["rpn"]["rpn_boxes"] / 2)
-    total_neg_bboxes = int(config["rpn"]["rpn_boxes"] / 2)
-    variances = torch.tensor(config["rpn"]["variances"]).to(device)
-    adaptive_ratio = config["rpn"]["adaptive_ratio"]
-    postive_th = config["rpn"]["positive_th"]
-    output_height, output_width = config["feature_map_shape"]
+    anchor_count = rpn_cfg["anchor_count"]
+    total_pos_bboxes = int(rpn_cfg["rpn_boxes"] / 2)
+    total_neg_bboxes = int(rpn_cfg["rpn_boxes"] / 2)
+    variances = torch.tensor(rpn_cfg["variances"]).to(device)
+    adaptive_ratio = rpn_cfg["adaptive_ratio"]
+    postive_th = rpn_cfg["positive_th"]
+    output_height, output_width = feature_map_shape
     anchors = anchors.to(device)
 
     iou_map = generate_iou_map(anchors, gt_boxes)
