+++ detection_vis_backend/train/utils.py
@@ -451,60 +451,6 @@ def get_metrics(metrics):
     return metrics_values
 
 
-def boxDecoder(yolohead_output, input_size, anchors_layer, num_class, scale=1., device='cuda:0'):
-    """ Decoder output from yolo head to boxes """ 
-
-    grid_size = yolohead_output.shape[1:4]
-    num_anchors_layer = len(anchors_layer)
-    grid_strides = torch.tensor(input_size, dtype=torch.float32).to(device) / torch.tensor(grid_size, dtype=torch.float32).to(device)
-    reshape_size = [yolohead_output.shape[0]] + list(grid_size) + [num_anchors_layer, 7+num_class]
-    pred_raw = yolohead_output.view(reshape_size)  
-    raw_xyz, raw_whd, raw_conf, raw_prob = torch.split(pred_raw, (3,3,1,num_class), dim=-1)
-
-    xyz_grid = torch.meshgrid(torch.arange(grid_size[0]).to(device), 
-                              torch.arange(grid_size[1]).to(device),
-                              torch.arange(grid_size[2]).to(device), indexing="ij") # Added indexing style
-    xyz_grid = torch.unsqueeze(torch.stack(xyz_grid, dim=-1).to(device), dim=3)
-    xyz_grid = xyz_grid.permute(1, 0, 2, 3, 4)
-    xyz_grid = xyz_grid.unsqueeze(0).repeat(yolohead_output.size(0), 1, 1, 1, num_anchors_layer, 1)
-
-    pred_xyz = ((torch.sigmoid(raw_xyz) * scale) - 0.5 * (scale - 1) + xyz_grid) * grid_strides
-
-    # Clipping values 
-    raw_whd = torch.clamp(raw_whd, 1e-12, 1e12)
-    
-    pred_whd = torch.exp(raw_whd) * torch.tensor(anchors_layer, dtype=torch.float32).to(device)
-    pred_xyzwhd = torch.cat([pred_xyz, pred_whd], dim=-1)
-
-    pred_conf = torch.sigmoid(raw_conf)
-    pred_prob = torch.sigmoid(raw_prob)
-    
-    return pred_raw, torch.cat([pred_xyzwhd, pred_conf, pred_prob], dim=-1)
-
-
-def extractYoloInfo(yolo_output_format_data):
-    """ Extract box, objectness, class from yolo output format data """
-    box = yolo_output_format_data[..., :6]
-    conf = yolo_output_format_data[..., 6:7]
-    category = yolo_output_format_data[..., 7:]
-    return box, conf, category
-
-
-def yolo1Loss(pred_box, gt_box, gt_conf, input_size, if_box_loss_scale=True):
-    """ loss function for box regression (based on YOLOV1) """
-    assert pred_box.shape == gt_box.shape
-    if if_box_loss_scale:
-        scale = 2.0 - 1.0 * gt_box[..., 3:4] * gt_box[..., 4:5] * gt_box[..., 5:6] /\
-                                    (input_size[0] * input_size[1] * input_size[2])
-    else:
-        scale = 1.0
-        
-    # YOLOv1 original loss function
-    giou_loss = gt_conf * scale * ((pred_box[..., :3] - gt_box[..., :3]).pow(2) + \
-                    (pred_box[..., 3:].sqrt() - gt_box[..., 3:].sqrt()).pow(2))
-    return giou_loss
-
-
 def iou3d(box_xyzwhd_1, box_xyzwhd_2, input_size):
     """ PyTorch version of 3D bounding box IOU calculation 
     Args:
@@ -572,57 +518,6 @@ def iou3d_np(box_xyzwhd_1, box_xyzwhd_2, input_size):
     return iou
 
 
-def focalLoss(raw_conf, pred_conf, gt_conf, pred_box, raw_boxes, input_size, iou_loss_threshold=0.5):
-    """ Calculate focal loss for objectness """
-    iou = iou3d(pred_box.unsqueeze(-2), raw_boxes.unsqueeze(1).unsqueeze(1).unsqueeze(1).unsqueeze(1), input_size)
-    max_iou, _ = iou.max(dim=-1)
-    max_iou = max_iou.unsqueeze(-1)
-
-    gt_conf_negative = (1.0 - gt_conf) * (max_iou < iou_loss_threshold).float()
-    conf_focal = (gt_conf - pred_conf).pow(2)
-    alpha = 0.01
-
-    focal_loss = conf_focal * (
-        gt_conf * F.binary_cross_entropy_with_logits(raw_conf, gt_conf, reduction='none')
-        +
-        alpha * gt_conf_negative * F.binary_cross_entropy_with_logits(raw_conf, gt_conf, reduction='none')
-    )
-    return focal_loss
-
-
-def categoryLoss(raw_category, pred_category, gt_category, gt_conf):
-    """ Category Cross Entropy loss """
-    category_loss = gt_conf * F.binary_cross_entropy_with_logits(input=raw_category, target=gt_category)
-    return category_loss
-
-
-def lossYolo(pred_raw, pred, label, raw_boxes, input_size, focal_loss_iou_threshold):
-    """ Calculate loss function of YOLO HEAD 
-    Args:
-        feature_stages      ->      3 different feature stages after YOLO HEAD
-                                    with shape [None, r, a, d, num_anchors, 7+num_class]
-        gt_stages           ->      3 different ground truth stages 
-                                    with shape [None, r, a, d, num_anchors, 7+num_class]"""
-    assert len(raw_boxes.shape) == 3
-    input_size = torch.tensor(input_size).float()
-    assert pred_raw.shape == label.shape
-    assert pred_raw.shape[0] == len(raw_boxes)
-    assert pred.shape == label.shape
-    assert pred.shape[0] == len(raw_boxes)
-    raw_box, raw_conf, raw_category = extractYoloInfo(pred_raw)
-    pred_box, pred_conf, pred_category = extractYoloInfo(pred)
-    gt_box, gt_conf, gt_category = extractYoloInfo(label)
-    giou_loss = yolo1Loss(pred_box, gt_box, gt_conf, input_size, \
-                            if_box_loss_scale=False)
-    focal_loss = focalLoss(raw_conf, pred_conf, gt_conf, pred_box, raw_boxes, \
-                            input_size, focal_loss_iou_threshold)
-    category_loss = categoryLoss(raw_category, pred_category, gt_category, gt_conf)
-    giou_total_loss = torch.mean(torch.sum(giou_loss, dim=[1, 2, 3, 4]))
-    conf_total_loss = torch.mean(torch.sum(focal_loss, dim=[1, 2, 3, 4]))
-    category_total_loss = torch.mean(torch.sum(category_loss, dim=[1, 2, 3, 4]))
-    return giou_total_loss, conf_total_loss, category_total_loss
-
-
 def yoloheadToPredictions(yolohead_output, conf_threshold=0.5):
     """ Transfer YOLO HEAD output to [:, 8], where 8 means
     [x, y, z, w, h, d, score, class_index]"""
@@ -670,118 +565,6 @@ def nms(bboxes, iou_threshold, input_size, sigma=0.3, method='nms'):
         else:
             best_bboxes = np.zeros([0, 8])
     return best_bboxes
-
-
-class Cont_Loss(nn.Module):
-    '''nn.Module warpper for focal loss'''
-    def __init__(self, win_size):
-        super(Cont_Loss, self).__init__()
-        self.win_size = win_size
-        self.MSE_loss = nn.MSELoss()
-
-    def forward(self, out, target):
-        loss = 0
-        for i in range(out.shape[0]):
-            batch_loss = 0
-            for j in range(self.win_size):
-                if j % 2 == 0:
-                    batch_loss = batch_loss + self.MSE_loss(out[i, :, j, :, :], target[i, :, j + 1, :, :])
-            loss = loss + batch_loss/(self.win_size/2)
-            # loss = loss + batch_loss
-        return loss
-
-
-def _neg_loss(pred, gt):
-    ''' Modified focal loss. Exactly the same as CornerNet.
-        Runs faster and costs a little bit more memory
-    Arguments:
-        pred (batch x c x h x w)
-        gt_regr (batch x c x h x w)
-    '''
-    balance_cof = 4
-    # focal_inds = pred.gt(1.4013e-45) * pred.lt(1-1.4013e-45)
-    pred = torch.clamp(pred, 1.4013e-45, 1)
-    fos = torch.sum(gt, 1)
-    pos_inds = gt.eq(1).float()
-    neg_inds = gt.lt(1).float()
-    fos_inds = torch.unsqueeze(fos.gt(0).float(), 1)
-    fos_inds = fos_inds.expand(-1, 3, -1, -1, -1)
-    neg_inds = neg_inds + (balance_cof - 1) * fos_inds * gt.eq(0)
-
-    neg_weights = torch.pow(1 - gt, 4)
-    loss = 0
-    # pos_loss = torch.log(pred) * torch.pow(1 - pred, 2) * pos_inds * focal_inds
-    # neg_loss = torch.log(1 - pred) * torch.pow(pred, 2) * neg_weights * neg_inds * focal_inds
-    pos_loss = torch.log(pred) * torch.pow(1 - pred, 2) * pos_inds * balance_cof
-    neg_loss = torch.log(1 - pred) * torch.pow(pred, 2) * neg_weights * neg_inds
-
-    num_pos  = pos_inds.float().sum()
-    pos_loss = pos_loss.sum()
-    neg_loss = neg_loss.sum()
-
-    if num_pos == 0:
-        loss = loss - neg_loss
-    else:
-        loss = loss - (pos_loss + neg_loss) / num_pos
-    return loss
-
-class FocalLoss_Neg(nn.Module):
-    '''nn.Module warpper for focal loss'''
-    def __init__(self):
-        super(FocalLoss_Neg, self).__init__()
-        self.neg_loss = _neg_loss
-
-    def forward(self, out, target):
-        return self.neg_loss(out, target)
-
-
-class FocalLoss_weight(nn.Module):
-    def __init__(self,weights,alpha,beta):
-        super(FocalLoss_weight,self).__init__()
-        self.weights = weights
-        self.alpha = alpha
-        self.beta = beta
-
-    def forward(self,input,target):
-        BCE_loss = F.binary_cross_entropy_with_logits(input, target, reduction='none')
-        pt = torch.exp(-BCE_loss) # prevents nans when probability 0
-        F_loss = self.weights * (1-pt)**self.alpha * BCE_loss
-        
-        return F_loss.sum()
-
-class CenterLoss(nn.Module):
-    def __init__(self):
-        super(CenterLoss,self).__init__()
-    
-    def forward(self,pred_map,target_map):
-
-        target_map= target_map.flatten()
-        idx = torch.nonzero(target_map)
-        target_map = ((target_map/8)+0.5)
-        pred_map = pred_map.flatten()
-        c_loss = F.binary_cross_entropy_with_logits(pred_map[idx], target_map[idx], reduction='none')
-        #c_loss = F.binary_cross_entropy_with_logits(pred_map[idx], target_map[idx], reduction='none')
-        return c_loss.sum()
-
-class L2Loss(nn.Module):
-    def __init__(self):
-        super(L2Loss,self).__init__()
-   
-    def forward(self,pred_o,target_o):
-        ep = 1e-10
-        target_s= target_o[:,0,::].flatten() # sine 
-        target_c= target_o[:,1,::].flatten() # cosine
-
-        idx = torch.nonzero(target_c)
-
-        
-        pred_s= pred_o[:,0,::].flatten()
-        pred_c= pred_o[:,1,::].flatten()
-
-        loss = (F.mse_loss(pred_s[idx],target_s[idx],reduction='sum')
-                +F.mse_loss(pred_c[idx],target_c[idx],reduction='sum'))
-
-        return loss.sum()
     
 
 def create_default(map_shape, kernel_window):
